Add CreateNewPrompt component
